ref: Branch selector into just a single component (#2614)
Refactor the branch selector on the bundles tab to it's own component and remove a lot of extra files that aren't needed.
